trying whole bunch of stuff to get puppeteer working on Vercel
