+++ apps/web/package.json
@@ -21,6 +21,8 @@
     "@nawadi/core": "workspace:^",
     "@next/mdx": "14.2.0-canary.60",
     "@react-google-maps/api": "^2.19.3",
+    "@react-pdf/renderer": "^3.4.2",
+    "@sparticuz/chromium": "^123.0.1",
     "@supabase/ssr": "^0.1.0",
     "@supabase/supabase-js": "^2.42.5",
     "@tanstack/react-table": "^8.16.0",
@@ -40,16 +42,20 @@
     "micromark": "^4.0.0",
     "next": "14.3.0-canary.44",
     "pdf-lib": "^1.17.1",
+    "pdfkit": "^0.15.0",
     "posthog-js": "^1.116.2",
     "posthog-node": "^4.0.0",
     "puppeteer": "^22.7.1",
+    "puppeteer-core": "^22.8.0",
     "qrcode.react": "^3.1.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-wrap-balancer": "^1.1.0",
     "remark-gfm": "^3.0.1",
+    "satori": "^0.10.13",
     "server-only": "^0.0.1",
     "slugify": "^1.6.6",
+    "svg-to-pdfkit": "^0.1.8",
     "tailwind-merge": "^2.2.1",
     "zod": "^3.22.4"
   },
@@ -61,8 +67,10 @@
     "@tailwindcss/typography": "^0.5.10",
     "@types/eslint": "^8.56.5",
     "@types/node": "^20",
+    "@types/pdfkit": "^0.13.4",
     "@types/react": "^18.3.1",
     "@types/react-dom": "^18.3.0",
+    "@types/svg-to-pdfkit": "^0.1.3",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
     "@typescript-eslint/parser": "^7.2.0",
     "autoprefixer": "^10.0.1",
@@ -73,6 +81,7 @@
     "postcss": "^8.4.35",
     "prettier": "^3.2.5",
     "prettier-plugin-tailwindcss": "^0.5.12",
+    "sharp": "^0.33.3",
     "tailwindcss": "^3.3.0",
     "typescript": "^5.4.5"
   }
