+++ apps/web/src/app/_components/certificate/certificate-raw.tsx
@@ -1,0 +1,133 @@
+import clsx from "clsx";
+import dayjs from "dayjs";
+import { QRCodeSVG } from "qrcode.react";
+import type { PropsWithChildren } from "react";
+import type { retrieveCertificateById } from "~/lib/nwd";
+import bg from "./diploma-print-bg.png";
+import "./printStyles.css";
+
+const DataField = ({
+  children,
+  className,
+}: PropsWithChildren<{ className?: string }>) => (
+  <span
+    className={clsx(
+      "absolute w-[52.5mm] h-[5.512mm] pt-[1.411mm] font-base text-[12pt]",
+      className,
+    )}
+  >
+    {children}
+  </span>
+);
+
+const Module = ({ value }: { value: string }) => (
+  <span className="flex flex-col w-[52.5mm]">
+    <span className="text-[9pt] font-medium leading-[10pt]">{value}</span>
+
+    <svg
+      width="100%"
+      height="1pt"
+      className="text-branding-light h-[1pt] w-full mt-[0.723mm]"
+    >
+      <rect width="100%" height="100%" fill="currentColor" />
+    </svg>
+
+    <svg
+      width="100%"
+      height="1pt"
+      className="text-branding-light h-[1pt] w-full mt-[2pt]"
+    >
+      <rect width="100%" height="100%" fill="currentColor" />
+    </svg>
+  </span>
+);
+
+export default function RawCertificate({
+  certificate,
+  advice,
+}: {
+  certificate: Awaited<ReturnType<typeof retrieveCertificateById>>;
+  advice: string;
+}) {
+  const uniqueCompletedModules = Array.from(
+    new Set(
+      certificate.completedCompetencies.map(
+        (competency) => competency.curriculum_competency.moduleId,
+      ),
+    ),
+  );
+
+  const modules = certificate.curriculum.modules.filter((module) =>
+    uniqueCompletedModules.includes(module.id),
+  );
+
+  return (
+    <div className="page" style={{ backgroundImage: `url(${bg.src})` }}>
+      <span className="text-[20pt] font-bold absolute pr-[2mm] left-[168mm] leading-[7.322mm] top-[13mm] w-[96mm] h-[6.322mm] text-right">
+        {certificate.gearType.title}
+      </span>
+
+      <span className="text-[11pt] font-base absolute pr-[2mm] left-[168mm] top-[19.889mm] leading-[5.111mm] w-[96mm] h-[5.111mm] text-right">
+        {certificate.program.title?.slice(0, -1).trim()}
+      </span>
+
+      <span className="text-[42pt] size-[12mm] text-right font-black text-branding-orange leading-[12mm] absolute left-[264mm] top-[13mm]">
+        {certificate.program.degree.rang}
+      </span>
+
+      <span className="text-[10pt] text-justify leading-[12pt] flex items-center justify-center absolute left-[23mm] top-[177mm] w-[71mm] h-[22mm]">
+        <span>{`${advice} Scan de QR-code voor meer informatie.`}</span>
+      </span>
+
+      <div className="size-[28mm] absolute left-[100mm] p-[3mm] top-[174mm]">
+        <QRCodeSVG
+          className="w-full h-full"
+          value={`https://www.nationaalwatersportdiploma.nl/diploma?nummer=${certificate.handle}&datum=${dayjs(certificate.issuedAt).format("YYYYMMDD")}`}
+        />
+      </div>
+
+      <p className="absolute w-[108mm] h-[5.512mm] left-[168mm] top-[49.948mm] pt-[1.411mm] font-bold text-[12pt]">
+        {[
+          certificate.student.firstName,
+          certificate.student.lastNamePrefix,
+          certificate.student.lastName,
+        ]
+          .filter(Boolean)
+          .join(" ")}
+      </p>
+
+      <DataField className="left-[168mm] top-[71.106mm]">
+        {dayjs(certificate.student.dateOfBirth).format("DD-MM-YYYY")}
+      </DataField>
+
+      <DataField className="left-[223.5mm] top-[71.106mm]">
+        {`${certificate.student.birthCity}${certificate.student.birthCountry.code !== "nl" ? ` (${certificate.student.birthCountry.name})` : ""}`}
+      </DataField>
+
+      <DataField className="left-[168mm] top-[92.264mm]">
+        {dayjs(certificate.issuedAt).format("DD-MM-YYYY")}
+      </DataField>
+
+      <DataField className="left-[223.5mm] top-[92.264mm]">
+        {certificate.handle}
+      </DataField>
+
+      {certificate.location.logoCertificate ? (
+        <span className="absolute left-[171mm] top-[117.833mm] w-[46.5mm] h-[29mm]">
+          {/* eslint-disable-next-line @next/next/no-img-element */}
+          <img
+            className="w-full h-full object-contain"
+            src={certificate.location.logoCertificate.url}
+            alt={certificate.location.logoCertificate.alt ?? ""}
+          />
+        </span>
+      ) : null}
+
+      <span className="absolute left-[20mm] top-[62.5mm] w-[108mm] h-auto grid gap-x-[3mm] gap-y-[6mm] grid-cols-2">
+        {modules.map((module) =>
+          module.title ? <Module key={module.id} value={module.title} /> : null,
+        )}
+      </span>
+    </div>
+  );
+}
