+++ apps/web/src/app/_components/certificate/diploma-print-bg.png
